+++ inst/application/ui.R
@@ -1,6 +1,9 @@
 library(DT)
 library(plotly)
+<<<<<<< HEAD
 library(shinyFiles)
+=======
+>>>>>>> upstream/master
 
 .readYamlConfig <- function() {
   config <- example_strategy_config()
@@ -106,10 +109,15 @@ ui <- fluidPage(
             column(
               12,
               br(),
+<<<<<<< HEAD
               # plotlyOutput('plot_3'),
               uiOutput('plot_3s'),
               # plotlyOutput('plot_4')
               uiOutput('factor_exposure')
+=======
+              plotlyOutput('plot_3'),
+              plotlyOutput('plot_4')
+>>>>>>> upstream/master
             )
           )
         ),