+++ tests/testthat/test_Simulation.R
@@ -9,7 +9,6 @@ sim$run()
 
 test_that("simulation produces expected results", {
   
-  
   test_summary <- filter(sim$getSimSummary())
   test_exposures <- filter(sim$getExposures())
   
@@ -26,24 +25,10 @@ test_that("simulation produces expected results", {
 
 
 # For easier testing, load up the feather file data into objects.
-all_dates <- seq.Date(from = as.Date("2019-01-02"), to = as.Date("2019-01-08"), by = "days")
-all_dates <- all_dates[!weekdays(all_dates) %in% c("Saturday", "Sunday")]
-test_input_data <-
-  lapply(all_dates,
-         function(x) {
-           read_feather(paste0("data/test_input/inputs/inputs_", format(x, "%Y%m%d"), ".feather")) %>%
-             mutate(date = x)
-         }) %>% bind_rows
 
-test_pricing_data <-
-  lapply(all_dates,
-         function(x) {
-           read_feather(paste0("data/test_input/pricing/pricing_", format(x, "%Y%m%d"), ".feather")) %>%
-             mutate(date = x)
-         }) %>% bind_rows
-
+test_input_data <- strand:::load_data_files("data/test_input/inputs")
+test_pricing_data <- strand:::load_data_files("data/test_input/pricing")
 test_secref_data <- read_feather("data/test_input/secref.feather")
-
 
 test_that("simulation produces same result when data supplied as objects", {
   
@@ -188,6 +173,170 @@ test_that("fill_rate_pct limits order filling", {
   summary_df <- sim$getSimSummary() %>% filter(strategy %in% "joint")
   expect_equal(summary_df$fill_rate_pct, 65)
   
+})
+
+
+# Slightly more complicated setup with 4 securities.
+#
+# Default setup is that 101 and 102 have positive alpha, 103 and 104 have
+# negative alpha.
+test_ids <- as.character(101:104)
+simple_input_data_2 <- filter(test_input_data, .data$id %in% !!test_ids) %>%
+  mutate(alpha_1 = case_when(id %in% "101" ~ 3,
+                             id %in% "102" ~ 2.5,
+                             id %in% "103" ~ -2.5,
+                             id %in% "104" ~ -3))
+
+# Prices are 1 for 101, 2, for 102, ..., etc.
+simple_pricing_data_2 <- filter(test_pricing_data, .data$id %in% !!test_ids) %>%
+  mutate(price_unadj = as.numeric(id) %% 100,
+         prior_close_unadj = price_unadj)
+
+simple_secref_data_2 <- filter(test_secref_data, .data$id %in% !!test_ids)
+
+test_that("force_trim_factor triggers trimming of large positions", {
+  
+  # Price of stock 101 goes from 1 to 1.5 on 1/2:
+  simple_pricing_data_2 <- 
+    simple_pricing_data_2 %>%
+    mutate(price_unadj = replace(price_unadj, id %in% 101 & date >= as.Date("2019-01-02"), 1.5),
+           prior_close_unadj = replace(prior_close_unadj, id %in% 101 & date >= as.Date("2019-01-03"), 1.5))
+  
+  
+  # Setup: long-short balanced. Max position 50%.
+  sim_config <- yaml::yaml.load_file("data/test_Simulation_simple.yaml")
+  sim_config$to <- "2019-01-04"
+  sim_config$strategies$strategy_1$ideal_short_weight <- 1
+  sim_config$strategies$strategy_1$position_limit_pct_lmv <- 50
+  sim_config$strategies$strategy_1$position_limit_pct_smv <- 50
+
+  sim <- Simulation$new(sim_config,
+                        raw_input_data = simple_input_data_2, 
+                        raw_pricing_data = simple_pricing_data_2,
+                        security_reference_data = simple_secref_data_2)
+  sim$run()
+  
+  # Investigate:
+  #
+  # sim$getSimDetail(strategy_name = "joint") %>%
+  #   select(sim_date, id, strategy, shares, start_price, end_price, end_nmv, end_shares, max_pos_lmv, max_pos_smv)
+
+  # Without force_trim_factor set, position in 101 remains at
+  # gmv of 375 despite max position of 250.
+  
+  expect_equal(
+    sim$getSimDetail(strategy_name = "joint", security_id = "101") %>%
+      filter(sim_date %in% as.Date("2019-01-03")) %>%
+      pull(end_gmv), 
+    375)
+  
+  # Now run with force_trim_factor set to 1.2.
+  sim_config$simulator$force_trim_factor <- 1.2
+  sim <- Simulation$new(sim_config,
+                        raw_input_data = simple_input_data_2, 
+                        raw_pricing_data = simple_pricing_data_2,
+                        security_reference_data = simple_secref_data_2)
+  sim$run()
+
+  # With the force_trim_factor setting of 1.2, a trade is generated to trim
+  # the position in 101 to 120% * 250 = 300.
+  expect_equal(
+    sim$getSimDetail(strategy_name = "joint", security_id = "101") %>%
+      filter(sim_date %in% as.Date("2019-01-03")) %>%
+      pull(end_gmv), 
+    300)
+  
+  # Finally, re-run but set average volume of 101 to 50 for 2019-01-03. With
+  # trading_limit_pct_adv set to 100, that means trimming will be limited to 50
+  # on 1/3. The remaining 25 is trimmed on 1/4.
+  simple_input_data_2 <- simple_input_data_2 %>%
+    mutate(average_volume = replace(average_volume, id %in% "101" & date %in% as.Date("2019-01-03"), 45))
+    
+  sim <- Simulation$new(sim_config,
+                        raw_input_data = simple_input_data_2, 
+                        raw_pricing_data = simple_pricing_data_2,
+                        security_reference_data = simple_secref_data_2)
+  sim$run()
+  
+  expect_equal(
+    sim$getSimDetail(strategy_name = "joint", security_id = "101") %>%
+      filter(sim_date %in% as.Date(c("2019-01-03", "2019-01-04"))) %>%
+      pull(end_gmv), 
+    c(330, 300))
+})
+
+
+test_that("force_exit_non_investable triggers exiting positions in non-investable stocks", {
+
+  # Here we go back to a 100% max position size so that at the end of 1/2 we
+  # have a portfolio that has one position long and one position short. 104 has
+  # the most negative alpha score, so we will have a short position in it at the
+  # end of 1/2.
+  #
+  # We define the universe to be stocks with an average_volume measure greater
+  # than 5000.
+  #
+  # We then set the average_volume to 4,000 for stock 104 for 2019-01-03
+  # onwards; for all other stocks we set average_volume to 10,000.
+  
+  simple_pricing_data_2 <- simple_pricing_data_2 %>%
+    mutate(price_unadj = replace(price_unadj, id %in% 101 & date >= as.Date("2019-01-02"), 1.5),
+           prior_close_unadj = replace(prior_close_unadj, id %in% 101 & date >= as.Date("2019-01-03"), 1.5))
+  
+  simple_input_data_2 <- simple_input_data_2 %>%
+    mutate(average_volume = if_else(id %in% 104 & date >= as.Date("2019-01-03"), 4000, 10000))
+  
+    
+  # Setup: long-short balanced. Max position 50%.
+  sim_config <- yaml::yaml.load_file("data/test_Simulation_simple.yaml")
+  sim_config$to <- "2019-01-07"
+  sim_config$strategies$strategy_1$ideal_short_weight <- 1
+  sim_config$strategies$strategy_1$position_limit_pct_lmv <- 100
+  sim_config$strategies$strategy_1$position_limit_pct_smv <- 100
+  sim_config$strategies$strategy_1$trading_limit_pct_adv <- 10
+  sim_config$simulator$force_exit_non_investable <- TRUE
+  sim_config$simulator$universe <- "average_volume >= 5000"
+  
+  
+  sim <- Simulation$new(sim_config,
+                        raw_input_data = simple_input_data_2, 
+                        raw_pricing_data = simple_pricing_data_2,
+                        security_reference_data = simple_secref_data_2)
+  sim$run()
+  
+  # Investigate:
+  #
+  # sim$getSimDetail(strategy_name = "joint") %>%
+  #   select(sim_date, id, strategy, shares, start_price, end_price, end_nmv, end_shares, max_pos_lmv, max_pos_smv)
+  
+  # On 2019-01-03, stock 104 falls outside of the universe due to its low
+  # average_volume measure. Since force_exit_non_investable is set, trades to
+  # exit 104 are added on 2019-01-03 and 2019-01-04. 10% ADV = 400 can be traded
+  # each day for 104. So at the end of the day on 2019-01-03 the position in 104
+  # has nmv of -100. At the end of 2019-01-04 the position in 104 is flat, while
+  # the position in 103 has taken its place and is up to size of 400. By the end
+  # of 2019-01-07, the position in 104 has been fully replaced by the position
+  # in 103.
+
+  # Check that 104 is exited at rate of max 400 / day.
+  expect_equal(
+    sim$getSimDetail(strategy_name = "joint", security_id = "104") %>%
+      pull(end_gmv), 
+    c(500, 100, 0, 0))
+
+  # Check the setting of the investable column for 104.
+  expect_equal(
+    sim$getSimDetail(strategy_name = "joint", security_id = "104") %>%
+      pull(investable), 
+    c(TRUE, FALSE, FALSE, FALSE))
+  
+  # Check that 104 is replaced by 103 (which has a price of 3). One day lag is
+  # because force exit is applied outside of the optimization.
+  expect_equal(
+    sim$getSimDetail(strategy_name = "joint", security_id = "103") %>%
+      pull(end_gmv), 
+    c(0, 0, 399, 501))
+
 })
 
 # Tests of summary functions